Allow to override build date
in order to make builds reproducible.
See https://reproducible-builds.org/ for why this is good
and https://reproducible-builds.org/specs/source-date-epoch/
for the definition of this variable.
This date call only works with GNU date.

Also use UTC to be independent of timezones.
And use ISO date format as the only proper way to unambiguously
represent a date.